+++ test/history.js
@@ -193,4 +193,4 @@ module.exports = function (opts) {
   })
 }
 
-if (!module.parent) { module.exports(require('../defaults')) }
+if (!module.parent) { module.exports({}) }
