+++ test/links.js
@@ -132,4 +132,4 @@ module.exports = function (opts) {
   })
 }
 
-if (!module.parent) { module.exports(require('../defaults')) }
+if (!module.parent) { module.exports({}) }
