+++ test/log.js
@@ -138,4 +138,4 @@ module.exports = function (opts) {
   })
 }
 
-if (!module.parent) { module.exports(require('../defaults')) }
+if (!module.parent) { module.exports({}) }
