+++ test/validation.js
@@ -169,4 +169,4 @@ module.exports = function (opts) {
   })
 }
 
-if (!module.parent) { module.exports(require('../defaults')) }
+if (!module.parent) { module.exports(require('../')) }
